requirements.txt
tensorflow
numpy
scikit-learn
matplotlib
streamlit
Pillow
opencv-python-headless